chore: bump the golang-dependencies group across 1 directory with 3 updates
Bumps the golang-dependencies group with 3 updates in the / directory: [gitlab.com/gitlab-org/api/client-go](https://gitlab.com/gitlab-org/api/client-go), [golang.org/x/oauth2](https://github.com/golang/oauth2) and [golang.org/x/sys](https://github.com/golang/sys).

Updates `gitlab.com/gitlab-org/api/client-go` from 0.160.1 to 1.8.1
- [Release notes](https://gitlab.com/gitlab-org/api/client-go/tags)
- [Changelog](https://gitlab.com/gitlab-org/api/client-go/blob/main/CHANGELOG.md)
- [Commits](https://gitlab.com/gitlab-org/api/client-go/compare/v0.160.1...v1.8.1)

Updates `golang.org/x/oauth2` from 0.33.0 to 0.34.0
- [Commits](https://github.com/golang/oauth2/compare/v0.33.0...v0.34.0)

Updates `golang.org/x/sys` from 0.38.0 to 0.39.0
- [Commits](https://github.com/golang/sys/compare/v0.38.0...v0.39.0)

---
updated-dependencies:
- dependency-name: gitlab.com/gitlab-org/api/client-go
  dependency-version: 1.8.1
  dependency-type: direct:production
  update-type: version-update:semver-major
  dependency-group: golang-dependencies
- dependency-name: golang.org/x/oauth2
  dependency-version: 0.34.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: golang-dependencies
- dependency-name: golang.org/x/sys
  dependency-version: 0.39.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
  dependency-group: golang-dependencies
...

Changed id variable from int to int64 due to
`cannot use id (variable of type int) as int64` errors.

+++ client/gitlab/client.go
@@ -11,9 +11,9 @@ import (
 
 // Client represents a GitLab client
 type Client interface {
-	CancelPipelineBuild(path string, id int) error
+	CancelPipelineBuild(path string, id int64) error
 	CreatePipeline(path string, options *gitlab.CreatePipelineOptions) (*gitlab.Pipeline, error)
-	GetPipelineVariables(path string, id int) ([]*gitlab.PipelineVariable, error)
+	GetPipelineVariables(path string, id int64) ([]*gitlab.PipelineVariable, error)
 	ProtectRepositoryBranches(
 		path string,
 		options *gitlab.ProtectRepositoryBranchesOptions,
@@ -51,7 +51,7 @@ func NewGitLabClient(accessToken string, baseURL string, dryRunMode bool) (Clien
 }
 
 // CancelPipelineBuild cancel a pipeline
-func (c *gitLabClient) CancelPipelineBuild(path string, id int) error {
+func (c *gitLabClient) CancelPipelineBuild(path string, id int64) error {
 	if c.dryRunMode {
 		msg := fmt.Sprintf("gitlab.CancelPipelineBuild: path=%s,id=%d",
 			path, id,
@@ -83,7 +83,7 @@ func (c *gitLabClient) CreatePipeline(
 // GetPipelineVariables get the pipeline variables
 func (c *gitLabClient) GetPipelineVariables(
 	path string,
-	id int,
+	id int64,
 ) ([]*gitlab.PipelineVariable, error) {
 	if c.dryRunMode {
 		msg := fmt.Sprintf("gitlab.GetPipelineVariables: path=%s,id=%d",
